FIR/UAST: commonize base UMethod
