+++ plugins/uast-kotlin/src/org/jetbrains/uast/kotlin/declarations/KotlinUMethod.kt
@@ -24,50 +24,19 @@ import org.jetbrains.kotlin.psi.psiUtil.getParentOfType
 import org.jetbrains.kotlin.utils.SmartList
 import org.jetbrains.kotlin.utils.addToStdlib.firstIsInstanceOrNull
 import org.jetbrains.uast.*
-import org.jetbrains.uast.java.internal.JavaUElementWithComments
 import org.jetbrains.uast.kotlin.*
 import org.jetbrains.uast.kotlin.psi.UastFakeLightMethod
 import org.jetbrains.uast.kotlin.psi.UastKotlinPsiParameter
 
 open class KotlinUMethod(
     psi: PsiMethod,
-    final override val sourcePsi: KtDeclaration?,
+    sourcePsi: KtDeclaration?,
     givenParent: UElement?
-) : KotlinAbstractUElement(givenParent), UMethodTypeSpecific, UAnchorOwner, JavaUElementWithComments, PsiMethod by psi {
-
-    constructor(psi: KtLightMethod, givenParent: UElement?) : this(psi, getKotlinMemberOrigin(psi), givenParent)
-
-    override val comments: List<UComment>
-        get() = super<KotlinAbstractUElement>.comments
-
-    override val psi: PsiMethod = unwrap<UMethod, PsiMethod>(psi)
-
-    override val javaPsi = psi
-
-    override fun getSourceElement() = sourcePsi ?: this
-
-    private val kotlinOrigin = getKotlinMemberOrigin(psi.originalElement) ?: sourcePsi
-
-    override fun getContainingFile(): PsiFile? {
-        kotlinOrigin?.containingFile?.let { return it }
-        return unwrapFakeFileForLightClass(psi.containingFile)
-    }
-
-    override fun getNameIdentifier() = UastLightIdentifier(psi, kotlinOrigin)
-
-    override val uAnnotations: List<UAnnotation> by lz {
-        psi.annotations
-            .mapNotNull { (it as? KtLightElement<*, *>)?.kotlinOrigin as? KtAnnotationEntry }
-            .map { baseResolveProviderService.baseKotlinConverter.convertAnnotation(it, this) }
-    }
-
-    private val receiver by lz {
-        when (sourcePsi) {
-            is KtCallableDeclaration -> sourcePsi
-            is KtPropertyAccessor -> sourcePsi.property
-            else -> null
-        }?.receiverTypeReference
-    }
+) : BaseKotlinUMethod(psi, sourcePsi, givenParent) {
+    constructor(
+        psi: KtLightMethod,
+        givenParent: UElement?
+    ) : this(psi, getKotlinMemberOrigin(psi), givenParent)
 
     override val uastParameters by lz {
 
@@ -78,51 +47,12 @@ open class KotlinUMethod(
         }
 
         val lightParams = psi.parameterList.parameters
-        val receiver = receiver ?: return@lz lightParams.map { KotlinUParameter(it, parameterOrigin(it), this) }
+        val receiver = receiverTypeReference ?: return@lz lightParams.map { KotlinUParameter(it, parameterOrigin(it), this) }
         val receiverLight = lightParams.firstOrNull() ?: return@lz emptyList<UParameter>()
         val uParameters = SmartList<UParameter>(KotlinReceiverUParameter(receiverLight, receiver, this))
         lightParams.drop(1).mapTo(uParameters) { KotlinUParameter(it, parameterOrigin(it), this) }
         uParameters
     }
-
-    override val uastAnchor by lazy {
-        KotlinUIdentifier(
-            nameIdentifier,
-            sourcePsi.let { sourcePsi ->
-                when (sourcePsi) {
-                    is PsiNameIdentifierOwner -> sourcePsi.nameIdentifier
-                    is KtObjectDeclaration -> sourcePsi.getObjectKeyword()
-                    is KtPropertyAccessor -> sourcePsi.namePlaceholder
-                    else -> sourcePsi?.navigationElement
-                }
-            },
-            this
-        )
-    }
-
-    override val uastBody by lz {
-        if (kotlinOrigin?.canAnalyze() != true) return@lz null // EA-137193
-        val bodyExpression = when (kotlinOrigin) {
-            is KtFunction -> kotlinOrigin.bodyExpression
-            is KtPropertyAccessor -> kotlinOrigin.bodyExpression
-            is KtProperty -> when {
-                psi is KtLightMethod && psi.isGetter -> kotlinOrigin.getter?.bodyExpression
-                psi is KtLightMethod && psi.isSetter -> kotlinOrigin.setter?.bodyExpression
-                else -> null
-            }
-            else -> null
-        } ?: return@lz null
-
-        wrapExpressionBody(this, bodyExpression)
-    }
-
-    override val returnTypeReference: UTypeReferenceExpression? by lz {
-        (sourcePsi as? KtCallableDeclaration)?.typeReference?.let {
-            KotlinUTypeReferenceExpression(it, this) { javaPsi.returnType ?: UastErrorType }
-        }
-    }
-
-    override fun equals(other: Any?) = other is KotlinUMethod && psi == other.psi
 
     companion object {
         fun create(psi: KtLightMethod, containingElement: UElement?): KotlinUMethod {
